Move util functions to Utils file
